One more Limit => int64

+++ txn/timelines.go
@@ -35,7 +35,7 @@ type GetTimeline_Hashtag struct {
 	MaxID     string   `query:"max_id"`
 	SinceID   string   `query:"since_id"`
 	MinID     string   `query:"min_id"`
-	Limit     int      `query:"limit"`
+	Limit     int64    `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
@@ -57,7 +57,7 @@ type GetTimeline_Home struct {
 	MaxID   string `query:"max_id"`
 	SinceID string `query:"since_id"`
 	MinID   string `query:"min_id"`
-	Limit   int    `query:"limit"`
+	Limit   int64  `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
@@ -80,7 +80,7 @@ type GetTimeline_List struct {
 	MaxID   string `query:"max_id"`
 	SinceID string `query:"since_id"`
 	MinID   string `query:"min_id"`
-	Limit   int    `query:"limit"`
+	Limit   int64  `query:"limit"`
 }
 
 // QueryPage implements the QueryPager interface, returning
